% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_matrix_profile.R
\name{compute_matrix_profile}
\alias{compute_matrix_profile}
\title{Read the ECG file generated by WFDB2MAT.}
\usage{
compute_matrix_profile(ecg_data, window_size = 200, n_workers = 4)
}
\description{
This function reads the tuple data.mat and data.hea that contains the signal and the header with information about
the signal.
}
