% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_model.R
\name{.check_contrast_profile_fit}
\alias{.check_contrast_profile_fit}
\title{Helper functions for checking arguments on fit and predict}
\usage{
.check_contrast_profile_fit(x)
}
\arguments{
\item{x}{An object of class \code{model_spec}.}
}
\description{
These functions are for developer use.

\code{.check_contrast_profile_fit()} checks that the model specification for fitting a
model contains a single value.

\code{.check_contrast_profile_predict()} checks that the argument value used for prediction is valid.
If called by \code{predict()}, it needs to be a single value. Multiple values are
allowed for \code{multi_predict()}.
}
\keyword{internal}
