% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_model.R
\name{contrast_model}
\alias{contrast_model}
\title{contrast_model}
\usage{
contrast_model(
  mode = "classification",
  num_shapelets = NULL,
  redundancy = NULL,
  engine = "contrast_profile"
)
}
\description{
contrast_model
}
