% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_all_files.R
\name{find_all_files}
\alias{find_all_files}
\title{Read the ECG file generated by WFDB2MAT.}
\usage{
find_all_files(
  path = getOption("target_dataset_path", default = "inst/extdata/physionet/"),
  long = FALSE
)
}
\description{
This function reads the tuple data.mat and data.hea that contains the signal and the header with information about
the signal.
}
