% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Train the model}
\usage{
fit_model(dataset)
}
\description{
This function reads the tuple data.mat and data.hea that contains the signal and the header with information about
the signal.
}
