% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_model.R, R/floss_model.R
\name{.check_floss_regime_threshold_fit}
\alias{.check_floss_regime_threshold_fit}
\title{Helper functions for checking the regime_threshold of FLOSS models}
\usage{
.check_floss_regime_threshold_fit(x)

.check_floss_regime_threshold_fit(x)
}
\arguments{
\item{x}{An object of class \code{model_spec}.}
}
\description{
These functions are for developer use.

\code{.check_floss_regime_threshold_fit()} checks that the model specification for fitting a
floss model contains a single value.

\code{.check_floss_regime_threshold_predict()} checks that the regime_threshold value used for prediction is valid.
If called by \code{predict()}, it needs to be a single value. Multiple values are
allowed for \code{multi_predict()}.

These functions are for developer use.

\code{.check_floss_regime_threshold_fit()} checks that the model specification for fitting a
floss model contains a single value.

\code{.check_floss_regime_threshold_predict()} checks that the regime_threshold value used for prediction is valid.
If called by \code{predict()}, it needs to be a single value. Multiple values are
allowed for \code{multi_predict()}.
}
\keyword{internal}
