% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floss_model.R
\name{floss_regime_model}
\alias{floss_regime_model}
\title{floss_regime_model}
\usage{
floss_regime_model(
  mode = "regression",
  window_size = NULL,
  time_constraint = NULL,
  mp_threshold = NULL,
  regime_threshold = NULL,
  regime_landmark = NULL,
  engine = "floss"
)
}
\description{
floss_regime_model
}
