% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floss_yardstick.R
\name{floss_score_pr}
\alias{floss_score_pr}
\title{Calculate the F-score for precision and recall}
\usage{
floss_score_pr(gtruth, reported, freq = 250, window = 10, beta = 4)
}
\arguments{
\item{gtruth}{A numeric vector or list of numeric vectors representing the ground truth values.}

\item{reported}{A numeric vector or list of numeric vectors representing the reported values.}

\item{freq}{The frequency of the data points (default is 250).}

\item{window}{The window size in seconds for matching the reported values with the ground truth values (default is 10).}

\item{beta}{The beta value for balancing precision and recall (default is 4).}
}
\value{
The F-score, a value between 0 and 1, representing the balance between precision and recall.
}
\description{
This function calculates the F-score, a measure of precision and recall, for a given ground truth and reported values.
}
\details{
This is a simpler version of the previous function. The recall delta is computed from freq * window.
Delta is then used to compute the overlap window between the ground truth and reported values. Half of delta is used
for earlier predictions, and full delta is used for later predictions. If this window contains several
reported values, just one is accounted as a true positive. If this window contains no reported values,
it is accounted as a false negative. The precision is computed as the number of using the previous tp and fn
while the false positives are computed as the number of reported values (all of them) that are not in any window.

The parameter \code{beta} is used to balance precision and recall. The default value is 4, which means that recall is
4 times more important than precision. If \code{beta} is 1, then precision and recall are equally important.
If \code{beta} is 0.25, then precision is 4 times more important than recall.
}
\examples{
# Calculate the F-score for a single ground truth and reported values
score_pr(c(1000, 2000, 3000, 4000), c(2000, 3000, 4000, 5000), window = 1) # 0.75

# Calculate the F-score for multiple ground truth and reported values
score_pr(list(c(1000, 2000, 3000, 4000), c(5000, 6000, 7000, 8000)),
  list(c(2000, 3000, 4000, 5000), c(6000, 7000, 8000, 9000)),
  window = 1
) # 0.75 0.75

}
\references{
For more information on the F-score, see: https://en.wikipedia.org/wiki/F1_score
}
