% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_idxs_split.R
\name{make_idxs_split}
\alias{make_idxs_split}
\title{Read the ECG file generated by WFDB2MAT.}
\usage{
make_idxs_split(size, prop = 0.75)
}
\description{
This function reads the tuple data.mat and data.hea that contains the signal and the header with information about
the signal.
}
