% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floss_model.R
\name{multi_predict._floss_regime_model}
\alias{multi_predict._floss_regime_model}
\title{Model predictions across many sub-models}
\usage{
\method{multi_predict}{`_floss_regime_model`}(
  object,
  new_data,
  type = NULL,
  regime_threshold = NULL,
  regime_landmark = NULL,
  ...
)
}
\arguments{
\item{object}{A \code{model_fit} object.}

\item{new_data}{A rectangular data object, such as a data frame.}

\item{type}{A single character value or \code{NULL}. Possible values are
\code{"numeric"}, \code{"class"}, \code{"prob"}, \code{"conf_int"}, \code{"pred_int"}, \code{"quantile"},
or \code{"raw"}. When \code{NULL}, \code{predict()} will choose an appropriate value
based on the model's mode.}

\item{...}{Optional arguments to pass to \code{predict.model_fit(type = "raw")}
such as \code{type}.}
}
\value{
A tibble with the same number of rows as the data being predicted.
There is a list-column named \code{.pred} that contains tibbles with
multiple rows per sub-model. Note that, within the tibbles, the column names
follow the usual standard based on prediction \code{type} (i.e. \code{.pred_class} for
\code{type = "class"} and so on).
}
\description{
For some models, predictions can be made on sub-models in the model object.
}
