% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ecg.R
\name{plot_ecg}
\alias{plot_ecg}
\title{Plots the ECG data imported using rdmat()}
\usage{
plot_ecg(data)
}
\arguments{
\item{data}{The object from rdmat()}
}
\value{
Returns invisibly the generated plot. Also plots using plotly.
}
\description{
Plots the ECG data imported using rdmat()
}
\examples{

\dontrun{
  \donttest{
     data <- read_ecg("data/a103l")
     plot_ecg(data)
  }
}

}
