% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ecg.R
\name{read_ecg}
\alias{read_ecg}
\title{Read the ECG file generated by WFDB2MAT.}
\usage{
read_ecg(filename, plot = FALSE)
}
\arguments{
\item{filename}{character vector. Path to the file data. Without the extension .mat. Must exist both .mat and .hea
in the path}

\item{plot}{a logical value. If the imported file will be plotted. Default is FALSE.}
}
\value{
Returns a \code{list} with the signals and the attributes of the file:
\itemize{
\item time: timestamp of each observation, in seconds
\item "signal": the value of the signal on each timestamp. The name of this object is the name of the signal, for
example 'II', 'PLETH'.
\itemize{
\item signal attributes "info":
\itemize{
\item baseline: numeric value
\item grain: numeric value
\item unit: char
}
}
\item dataset attributes "info":
\itemize{
\item alarm: type of alarm, for ex: "Asystole".
\item true: if the alarm is true or not, for ex: "False".
\item filename: the name of the original file, without the extension.
\item frequency: the frequency of the observations, in Hz.
}
}
}
\description{
This function reads the tuple data.mat and data.hea that contains the signal and the header with information about
the signal.
}
\section{Acknowledgment}{
This is a R port from the Matlab's code written by Ikaro Silva, 2014, Version 1.1
}

\section{Known limitations}{
This function currently does support several of the features described in the WFDB record format (such as
multiresolution signals):
http://www.physionet.org/physiotools/wag/header-5.htm
If you are not sure that the record (or database format) you are reading is supported, you can do an integrity
check by comparing the output with RDSAMP.

if(sum(abs(signal-signal2)) !=0);
stop('Record not compatible with RDMAT');
end

These tools are available at: https://archive.physionet.org/physiotools/binaries/
}

\examples{
\dontrun{
signal <- read_ecg("data/a103l")
}

}
