% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_model.R
\name{test_model}
\alias{test_model}
\title{Test the model}
\usage{
test_model(model, dataset)
}
\description{
This function reads the tuple data.mat and data.hea that contains the signal and the header with information about
the signal.
}
