% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{contrast}
\alias{contrast}
\title{Contrast Profile}
\usage{
contrast(
  negative,
  positive,
  window_size,
  positive_matrix = NULL,
  exclusion_zone = 0.5,
  distance = "euclidean",
  progress = FALSE
)
}
\arguments{
\item{distance}{(\code{mpx()} only) A string. Currently accepts \code{euclidean} and \code{pearson}. Defaults to \code{euclidean}.}
}
\description{
Contrast Profile
}
\details{
\subsection{mpx}{

This algorithm was developed apart from the main Matrix Profile branch that relies on Fast Fourier Transform (FFT) at
least in one part of the process. This algorithm doesn't use FFT at all and is several times faster. It also relies
on Ogita's work for better precision computing mean and standard deviation (part of the process).
}
}
\section{This document}{
Last updated on 2024-02-22 using R version 4.3.2.
}

\seealso{
\code{mpxab()} for the forward and reverse join-similarity.
}
